develop "User & Usermanager Class" and change data type of some variables

+++ Include/Model/Users/Hash.h
@@ -8,6 +8,6 @@ class Hash
 {
 private:
 public:
-    int ConvertToHash(const string &);
+    size_t ConvertToHash(const string &);
 };
 #endif HASH_H
